requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
dbus-python==1.3.2
evdev==1.6.0
idna==3.4
python-mpv==1.0.1
python-xlib==0.31
six==1.16.0
urllib3==1.26.12
